refactor: apply agreed DDD package structure and service split
- move user package under com.gachi.be.domain.user (api/controller, dto/response)
- keep file as top-level shared package and reorganize api/controller, dto/response
- split S3 file service into interface + impl (S3FileService, S3FileServiceImpl)
- update imports/package declarations and pass spotlessCheck/test
