+++ src/main/java/com/gachi/be/file/dto/response/S3UploadResponse.java
@@ -1,3 +1,3 @@
-package com.gachi.be.file.dto;
+package com.gachi.be.file.dto.response;
 
 public record S3UploadResponse(String key, String url) {}
